refactor: 키워드 검색시 옵션을 받도록 구현
- 옵션에서 CategoryGroupCode를 통해 원하는 그룹 필터링 가능(음식점)
- 모호한 파일명 수정
